% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenberg_authors.R
\docType{data}
\name{gutenberg_authors}
\alias{gutenberg_authors}
\title{Metadata about Project Gutenberg authors}
\format{
A tbl_df (see \code{\link[tibble:tibble]{tibble::tibble()}}) with one row for each author, with
the columns
\describe{
\item{gutenberg_author_id}{Unique identifier for the author that can be
used to join with the \code{gutenberg_metadata} dataset (coming soon to this
package)}
\item{author}{The \code{agent_name} field from the original metadata}
\item{birthdate}{Year of birth}
\item{deathdate}{Year of death}
}
}
\usage{
gutenberg_authors
}
\description{
Data frame with metadata about each author of a Project Gutenberg work.
}
\details{
To find the date on which this metadata was last updated, see
\code{attr(gutenberg_authors, "date_updated")}.
}
\examples{

# date last updated
attr(gutenberg_authors, "date_updated")
}
\keyword{datasets}
