% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenbergr.data-package.R
\docType{package}
\name{gutenbergr.data-package}
\alias{gutenbergr.data}
\alias{gutenbergr.data-package}
\title{gutenbergr.data: Project Gutenberg Metadata}
\description{
Frequently updated metadata about the Project Gutenberg collection \url{https://www.gutenberg.org/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonthegeek/gutenbergr.data}
  \item \url{https://jonthegeek.github.io/gutenbergr.data/}
  \item Report bugs at \url{https://github.com/jonthegeek/gutenbergr.data/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
